more moving files, fix references
