+++ frontend/src/App.vue
@@ -90,11 +90,11 @@ import _ from 'lodash';
 import Vue from 'vue';
 import Events from './events';
 import MetaMaskOnboarding from '@metamask/onboarding';
-import BigButton from './components/BigButton.vue';
-import SmallButton from './components/SmallButton.vue';
-import NavBar from './components/layout/navigation/NavBar.vue';
-import CharacterBar from './components/layout/CharacterBar.vue';
-import WeaponRowGrid from './components/layout/WeaponRowGrid.vue';
+import BigButton from './components/layout/buttons/BigButton.vue';
+import SmallButton from './components/layout/buttons/SmallButton.vue';
+import NavBar from './components/layout/navigation/bars/NavBar.vue';
+import CharacterBar from './components/layout/bars/CharacterBar.vue';
+import WeaponRowGrid from './components/layout/grids/WeaponRowGrid.vue';
 import { apiUrl } from './utils/common';
 import i18n from './i18n';
 import { getConfigValue } from './contracts';
