+++ frontend/src/components/blacksmith/RightMenu.vue
