+++ frontend/src/components/character/CharacterList.vue
@@ -163,7 +163,7 @@
 <script>
 import {mapActions, mapGetters, mapState} from 'vuex';
 import {getCharacterArt} from '@/character-arts-placeholder';
-import CharacterArt from '../CharacterArt.vue';
+import CharacterArt from './CharacterArt.vue';
 import NftOptionsDropdown from '../nft/NftOptionsDropdown.vue';
 import {getCleanName, isProfaneIsh} from '@/rename-censor';
 import Events from '@/events';
