+++ frontend/src/components/layout/bars/CharacterBar.vue
@@ -20,7 +20,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import CharacterDisplay from '../character/CharacterDisplay.vue';
+import CharacterDisplay from '../../character/CharacterDisplay.vue';
 import '@/mixins/general';
