+++ frontend/src/components/layout/bars/SmallBar.vue
