+++ frontend/src/components/layout/bars/StaminaBar.vue
