+++ frontend/src/components/layout/buttons/SmallButton.vue
