+++ frontend/src/components/layout/navigation/bars/PvPNavBar.vue
@@ -2,19 +2,19 @@
   <div class="wrapper">
     <nav class="defaultNav">
       <button @click="setTab(0)" :class="tabNumber === 0 && 'active'">
-        <img src="../../../assets/swordsIcon.svg"/>
+        <img src="../../../../assets/swordsIcon.svg"/>
         <span>
           {{$t('pvp.arena')}}
         </span>
       </button>
       <button @click="setTab(1)" :class="tabNumber === 1 && 'active'">
-        <img src="../../../assets/crownIcon.svg" />
+        <img src="../../../../assets/crownIcon.svg" />
         <span>
           {{$t('pvp.leaderboard')}}
         </span>
       </button>
       <button @click="setTab(2)" :class="tabNumber === 2 && 'active'">
-        <img src="../../../assets/crownIcon.svg" />
+        <img src="../../../../assets/crownIcon.svg" />
         <div class="rewardsButtonInner">
           {{$t('pvp.rewards')}}
           <BIcon icon="exclamation-circle-fill" v-if="hasRewards" class="rewardsIcon"/>
