+++ frontend/src/components/loader/SkeletonLoader.vue
