+++ frontend/src/views/Blacksmith.vue
@@ -610,15 +610,15 @@
 
 <script lang='ts'>
 import BN from 'bignumber.js';
-import WeaponGrid from '../components/layout/WeaponGrid.vue';
-import RightMenu from '../components/layout/RightMenu.vue';
-import BigButton from '../components/BigButton.vue';
+import WeaponGrid from '../components/layout/grids/WeaponGrid.vue';
+import RightMenu from '../components/blacksmith/RightMenu.vue';
+import BigButton from '../components/layout/buttons/BigButton.vue';
 import { getWeaponArt } from '@/weapon-arts-placeholder';
 import { getWeaponRarity } from '@/weapon-element';
 import { getCleanName } from '@/rename-censor';
 import Vue from 'vue';
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex';
-import WeaponIcon from '../components/WeaponIcon.vue';
+import WeaponIcon from '../components/blacksmith/WeaponIcon.vue';
 import { BModal } from 'bootstrap-vue';
 import NftList from '@/components/nft/NftList.vue';
 import { Contracts, IState } from '@/interfaces';
