+++ frontend/src/views/Plaza.vue
@@ -335,7 +335,7 @@
 
 <script lang='ts'>
 import BN from 'bignumber.js';
-import BigButton from '../components/BigButton.vue';
+import BigButton from '../components/layout/buttons/BigButton.vue';
 import CharacterList from '../components/character/CharacterList.vue';
 import Hint from '../components/Hint.vue';
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex';
